KeyManager: add negotiation_delay and protocol jitter
A design bug in the protocol fsm caused Vita nodes to enter a state where the
protocol failed to progress. This commit negates this failure mode by
introducing a negotiation_delay (same duration as negotiation_ttl) that a
protocol fsm will spend idle after a negotiation was aborted because it
exceeded negotiation_ttl, and adding random jitter to both rekey_timeout and
negotiation_delay to create a time window where only either fsm is active.

This is a temporary hack: the protocol fsm is misguidedly designed to support
two simultaneously active parties--a property which is irrelevant in retrospect
as this will (almost) never happen in practice--in order to avoid the
complexity of proper multiplexing of active/passive protocol peers.