vita: add nat-ip option for public interfaces
This option is used by the KeyManager app to override the (potentially NAT-ed)
node IP as the node’s network identity. The network identity is included in the
AKE protocol’s prologue, and thus has to match what remote nodes see in order
for authentication to succeed. In cases where public interfaces are behind NAT
this option must be set to the translated public IP address of the node so that
the AKE prologue matches the prologue of remote nodes.