connectx.lua: Add CQE compression, disabled by default, check capability
CQE compression enables the hardware to reduce the number of full CQEs
getting posted to a CQ, which should provide a boost in performance
with certain traffic loads.  Essentilly, the algorithm generates a
single full CQE for a sequence of CQEs that differ only in the
checksum and the packet length and adds a compact data structure that
only stores the differing values from which the original CQEs can be
reconstructed.

The implementation was reverse-engineered from the mlx5 Linux driver
as the ConnectX4 PRM provides no documentation on the feature.

The feature can only be enabled if both, cqe_compression and
mini_cqe_resp_stride_index are available. The latter provides the WQ
consumer counter directly in the mini CQE rather than deriving it from
the CQE that starts the compression block.

# Conflicts:
#	src/apps/mellanox/connectx.lua