ipfix probe: cleanup code, optimize operation without software rss
This reorganizes the program.ipfix.lib module and optimizes the case
of a single embedded export instance and no flow director classes.

In this case, we do not create the software rss app and simply Join
the inputs to be then processed directly by the ipfix app.
NB: the ipfix app has to take care of extracting packet metadata.