apps.mellanox.connectx: alloc_pages(): remove assertion
Ignore 0 and negative values in alloc_pages(). This assumes that the
function is called with the result of query_pages() for which these
values are legal (a negative value indicates that pages can be
recalimed by the driver).  In some versions of the firmware on
ConnectX7, query_pages() for "regular" pages during initialization
returns a negative value that would trigger the assertion.

Allocation of "regular" pages is not part of the card initialization
according to the PRM, i.e. this particular problem could also be
avoided by removing that call. However, this commit seems more robust.

Note that page reclaims are already handled by the event queue through
reclaim notifications.