lib.ipc.shmem.iftable_mib: add ifIndex to the shmem segment
The SNMP agent only generates objects that are included explicitly in
the shmem segment. Even though the ifIndex value is created by the
agent itself, the OID is not instantiated without ifIndex being
present in the shmem segment.

Background: SMIv2 (RFC2578) specifies

   Objects which are both specified in the INDEX clause of a
   conceptual row and also columnar objects of the same conceptual row
   are termed auxiliary objects. The MAX-ACCESS clause for auxiliary
   objects is "not-accessible"

The current implementation assumed that ifIndex was of this type and
need not be instantiated by the SNMP agent. However, the ifTable MIB
predates this specification and declares MAX-ACCESS as
"read-only". Therefore, even though it is rare, an SNMP manager can
expect that this object exists.